Added "Freeze Weather" option
Freeze Weather option can be found under the "Weather" sub-menu. It
prevents the weather from changing if it's activated and you select one
of the weather options. If you have changed the weather before
activating the freeze and the weather has returned to normal, if you
then activate the setting it will freeze the weather at the last chosen
option.