Update vehicles.cpp
Set custom tag string via onscreen keyboard input.

This is more an excuse to share my code for using the ingame keyboard.
Abuse it. Seriously. It took me way too long to make it work and I cannot come up with a better use for it so have at it.