Improvements to chauffeur and Mr. Monkey behavior.
- Moved blipCoords to global scope so we don't set it every time.
There's no point to setting it more than once because we only call the
drive function once.

- Checks to see if player reached blip. I don't think it's a good idea
to check if a blip disappeared because, unlike the teleport function, a
player can change the blip location in between setting his destination
and actually getting there. We set the destination once, and then we
perform a proximity check based on the tolerance to determine if the
player reached his destination. The blip can change during the ride, but
the destination will remain the same until a new chauffeur is called.

- Renamed the menu option to indicate that the chauffeur can get
canceled.

- Removed canceling chauffer when leaving the car. It's a little buggy
right now in that it detects me as having left the car immediately upon
me calling the chauffeur.

- Updated the vehicle (again!) to a gold Armored Kuruma. Figured it'll
be more useful in case the driver rolls through some hostile territory.
We wouldn't want our player getting capped along the way.

- Updated Mr. Monkey's behavior. Now when we reach our destination, he
gets out of the vehicle and wanders around.