More updates to chauffeur
- Trying AI::TASK_VEHICLE_DRIVE_TO_COORD again. Looks like we pass the
function the vehicle entity, might be used for AI vehicle handling. In
either case, the AI driving seems to work a lot better using this
function. Drive mode of 4 is working really well so far.

- Updated is_player_at_blip, which is now based off a 2D Euclidean
distance. We can incorporate the Z coordinate at some point, but I don't
really care if Mr. Monkey drops us off at the bottom of a mountain or
tall building.

- Chauffeur tolerance is pretty tight right now. The guy gets pretty
close to being right on top of the blip. Feel free to change it around
if we want him to be a little more lax on where he drops you off.