Bodyguard fixes
Allows a maximum of 7 bodyguards, since the maximum group size is 8.
Trying to add bodyguards beyond the 7 causes the new ones to create
their own group and treat each other as neutral (hostile when fired
upon).

Bodyguard cleanup currently only checks for dead bodyguards. Not sure if
we need to be checking for anything else. I have a feeling
PED::_IS_PED_DEAD returns true if the PED is invalid for any reason
(i.e., we got too far and he despawned) but I haven't verified.