Minor updates and fixes
- Sorted out most of the the dial colours using the wrong colour
(Vermillion was using Orange for example and vice versa). Purple is
still broken however (don't know the ID for it. It seems to be using
Taxi Yellow or Choco orange)

- Modified the amount of items in the database.cpp so it now saves all
of Benny's mods (including Dash colour I believe) but not dial colour
(not sure why - will need a further look).

- Added skin options for Revolver (Normal, VIP and Bodyguard)